Split out reolink tests
